+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Script to create graph visualization of calculated data documents from asm json files
 - Details of parser requirements to docs
 - Add Agilent Gen5 Image Adapter
+- Add Methodical Mind adapter
 
 ### Fixed